introduce other c++ features. prepare to link with std lib

+++ Kernel/inc/stdcxx/types.h
@@ -15,7 +15,7 @@ using word  = unsigned short;
 using dword = unsigned int;
 using qword = unsigned long long;
 
-using size_t = unsigned long long;
+using size_t = long unsigned int;
 
 // using le8_t  = uint8_t __attribute__((bitwise));
 // using le16_t = uint16_t __attribute__((bitwise));
